Make ZeroVec deref to ZeroSlice
