+++ utils/zerovec/src/zerovec/mod.rs
@@ -14,6 +14,7 @@ use alloc::vec::Vec;
 use core::cmp::{Ord, Ordering, PartialOrd};
 use core::fmt;
 use core::iter::FromIterator;
+use core::ops::Deref;
 
 /// A zero-copy vector for fixed-width types.
 ///
@@ -25,6 +26,8 @@ use core::iter::FromIterator;
 /// see [`VarZeroVec`](crate::VarZeroVec).
 ///
 /// Typically, the zero-copy equivalent of a `Vec<T>` will simply be `ZeroVec<'a, T>`.
+///
+/// Most of the methods on `ZeroVec<'a, T>` come from its [`Deref`] implementation to [`ZeroSlice<T>`](ZeroSlice)
 ///
 /// # How it Works
 ///
@@ -94,6 +97,17 @@ where
     /// assert_eq!(bytes, zerovec.as_bytes());
     /// ```
     Borrowed(&'a [T::ULE]),
+}
+
+impl<'a, T: AsULE> Deref for ZeroVec<'a, T> {
+    type Target = ZeroSlice<T>;
+    fn deref(&self) -> &Self::Target {
+        let slice = match self {
+            ZeroVec::Owned(vec) => &**vec,
+            ZeroVec::Borrowed(slice) => *slice,
+        };
+        ZeroSlice::from_slice(slice)
+    }
 }
 
 impl<T> fmt::Debug for ZeroVec<'_, T>
@@ -195,79 +209,6 @@ where
     pub fn parse_byte_slice(bytes: &'a [u8]) -> Result<Self, ZeroVecError> {
         let slice: &'a [T::ULE] = T::ULE::parse_byte_slice(bytes)?;
         Ok(Self::Borrowed(slice))
-    }
-
-    /// Returns a `ZeroVec<T>` as its underlying `&[u8]` byte buffer representation.
-    ///
-    /// Useful for serialization.
-    ///
-    /// # Example
-    ///
-    /// ```
-    /// use zerovec::ZeroVec;
-    ///
-    /// // The little-endian bytes correspond to the numbers on the following line.
-    /// let bytes: &[u8] = &[0xD3, 0x00, 0x19, 0x01, 0xA5, 0x01, 0xCD, 0x01];
-    /// let nums: &[u16] = &[211, 281, 421, 461];
-    ///
-    /// let zerovec = ZeroVec::alloc_from_slice(nums);
-    ///
-    /// assert_eq!(bytes, zerovec.as_bytes());
-    /// ```
-    #[inline]
-    pub fn as_bytes(&self) -> &[u8] {
-        T::ULE::as_byte_slice(self.as_slice())
-    }
-
-    /// Dereferences this `ZeroVec<T>` as `&[T::ULE]`. Most other functions on `ZeroVec<T>` use
-    /// this function as a building block.
-    #[inline]
-    pub fn as_slice(&self) -> &[T::ULE] {
-        match self {
-            Self::Owned(vec) => vec.as_slice(),
-            Self::Borrowed(slice) => slice,
-        }
-    }
-
-    /// Returns the number of elements in this `ZeroVec<T>`.
-    ///
-    /// # Example
-    ///
-    /// ```
-    /// use zerovec::ZeroVec;
-    /// use zerovec::ule::AsULE;
-    ///
-    /// let bytes: &[u8] = &[0xD3, 0x00, 0x19, 0x01, 0xA5, 0x01, 0xCD, 0x01];
-    /// let zerovec: ZeroVec<u16> = ZeroVec::parse_byte_slice(bytes).expect("infallible");
-    ///
-    /// assert_eq!(4, zerovec.len());
-    /// assert_eq!(
-    ///     bytes.len(),
-    ///     zerovec.len() * std::mem::size_of::<<u16 as AsULE>::ULE>()
-    /// );
-    /// ```
-    #[inline]
-    pub fn len(&self) -> usize {
-        self.as_slice().len()
-    }
-
-    /// Returns whether the vec is empty.
-    ///
-    /// # Example
-    ///
-    /// ```
-    /// use zerovec::ZeroVec;
-    ///
-    /// let bytes: &[u8] = &[0xD3, 0x00, 0x19, 0x01, 0xA5, 0x01, 0xCD, 0x01];
-    /// let zerovec: ZeroVec<u16> = ZeroVec::parse_byte_slice(bytes).expect("infallible");
-    /// assert!(!zerovec.is_empty());
-    ///
-    /// let emptyvec: ZeroVec<u16> = ZeroVec::parse_byte_slice(&[]).expect("infallible");
-    /// assert!(emptyvec.is_empty());
-    /// ```
-    #[inline]
-    pub fn is_empty(&self) -> bool {
-        self.as_slice().is_empty()
     }
 
     /// Converts a `ZeroVec<T>` into a `ZeroVec<u8>`, retaining the current ownership model.
@@ -566,84 +507,6 @@ impl<T> ZeroVec<'_, T>
 where
     T: AsULE,
 {
-    /// Gets the element at the specified index. Returns None if out of range.
-    ///
-    /// # Example
-    ///
-    /// ```
-    /// use zerovec::ZeroVec;
-    ///
-    /// let bytes: &[u8] = &[0xD3, 0x00, 0x19, 0x01, 0xA5, 0x01, 0xCD, 0x01];
-    /// let zerovec: ZeroVec<u16> = ZeroVec::parse_byte_slice(bytes).expect("infallible");
-    ///
-    /// assert_eq!(zerovec.get(2), Some(421));
-    /// assert_eq!(zerovec.get(4), None);
-    /// ```
-    #[inline]
-    pub fn get(&self, index: usize) -> Option<T> {
-        self.as_slice().get(index).copied().map(T::from_unaligned)
-    }
-
-    pub(crate) fn get_ule_ref(&self, index: usize) -> Option<&T::ULE> {
-        self.as_slice().get(index)
-    }
-
-    /// Gets the first element. Returns None if empty.
-    ///
-    /// # Example
-    ///
-    /// ```
-    /// use zerovec::ZeroVec;
-    ///
-    /// let bytes: &[u8] = &[0xD3, 0x00, 0x19, 0x01, 0xA5, 0x01, 0xCD, 0x01];
-    /// let zerovec: ZeroVec<u16> = ZeroVec::parse_byte_slice(bytes).expect("infallible");
-    ///
-    /// assert_eq!(zerovec.first(), Some(211));
-    /// ```
-    #[inline]
-    pub fn first(&self) -> Option<T> {
-        self.as_slice().first().copied().map(T::from_unaligned)
-    }
-
-    /// Gets the last element. Returns None if empty.
-    ///
-    /// # Example
-    ///
-    /// ```
-    /// use zerovec::ZeroVec;
-    ///
-    /// let bytes: &[u8] = &[0xD3, 0x00, 0x19, 0x01, 0xA5, 0x01, 0xCD, 0x01];
-    /// let zerovec: ZeroVec<u16> = ZeroVec::parse_byte_slice(bytes).expect("infallible");
-    ///
-    /// assert_eq!(zerovec.last(), Some(461));
-    /// ```
-    #[inline]
-    pub fn last(&self) -> Option<T> {
-        self.as_slice().last().copied().map(T::from_unaligned)
-    }
-
-    /// Gets an iterator over the elements.
-    ///
-    /// # Example
-    ///
-    /// ```
-    /// use zerovec::ZeroVec;
-    ///
-    /// let bytes: &[u8] = &[0xD3, 0x00, 0x19, 0x01, 0xA5, 0x01, 0xCD, 0x01];
-    /// let zerovec: ZeroVec<u16> = ZeroVec::parse_byte_slice(bytes).expect("infallible");
-    /// let mut it = zerovec.iter();
-    ///
-    /// assert_eq!(it.next(), Some(211));
-    /// assert_eq!(it.next(), Some(281));
-    /// assert_eq!(it.next(), Some(421));
-    /// assert_eq!(it.next(), Some(461));
-    /// assert_eq!(it.next(), None);
-    /// ```
-    #[inline]
-    pub fn iter(&self) -> impl Iterator<Item = T> + '_ {
-        self.as_slice().iter().copied().map(T::from_unaligned)
-    }
-
     /// Mutates each element according to a given function, meant to be
     /// a more convenient version of calling `.iter_mut()` on
     /// [`ZeroVec::to_mut()`] which serves fewer use cases.
@@ -757,33 +620,6 @@ where
                 self.to_mut()
             }
         }
-    }
-}
-
-impl<T> ZeroVec<'_, T>
-where
-    T: AsULE + Ord,
-{
-    /// Binary searches a sorted `ZeroVec<T>` for the given element. For more information, see
-    /// the primitive function [`binary_search`].
-    ///
-    /// # Example
-    ///
-    /// ```
-    /// use zerovec::ZeroVec;
-    ///
-    /// let bytes: &[u8] = &[0xD3, 0x00, 0x19, 0x01, 0xA5, 0x01, 0xCD, 0x01];
-    /// let zerovec: ZeroVec<u16> = ZeroVec::parse_byte_slice(bytes).expect("infallible");
-    ///
-    /// assert_eq!(zerovec.binary_search(&281), Ok(1));
-    /// assert_eq!(zerovec.binary_search(&282), Err(2));
-    /// ```
-    ///
-    /// [`binary_search`]: https://doc.rust-lang.org/std/primitive.slice.html#method.binary_search
-    #[inline]
-    pub fn binary_search(&self, x: &T) -> Result<usize, usize> {
-        self.as_slice()
-            .binary_search_by(|probe| T::from_unaligned(*probe).cmp(x))
     }
 }
